% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{batcheffect.removal}
\alias{batcheffect.removal}
\title{Batch-Effect Removal}
\usage{
batcheffect.removal(amatrix, close_th = 0.01, cells_th = 0.05)
}
\arguments{
\item{amatrix}{matrix, each row corresponds to a cell, whose columns depict cluster membership of its neighbours; as ruturned by \code{find_ann()} module.}

\item{close_th}{numeric, [0,1] percentage of proximity between two genes.
Two genes are considered close when \code{(v2 - v1)/v1 < close_thresh}.}

\item{cells_th}{numeric, [0,1] percentage of cells which need to be close for one of the genes to be removed.}
}
\value{
numeric matrix where expression values are replaced with ranks.
}
\description{
Minimise batch effect by representig normalized expression values with suitable ranks
}
\details{
Minimising batch effect relies on an assumption that relative ordering of genes with
regards to their expressions are preserved across batches.
}
