% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster.cells}
\alias{cluster.cells}
\title{Clustering of samples}
\usage{
cluster.cells(data, sp.samples, default = TRUE, minClusterSize = 20,
  deepSplit = 3, conf = 0.75, k_nn = 10)
}
\arguments{
\item{data}{normalized expression matrix as returned by \code{matrix.subset()}.
The matrix columns represent the subset of genes obtained from \code{pc_genes} module.
Subsetting the return matrix by \code{matrix.subset()} module with samples from primary clusters and PCA selected genes.}

\item{sp.samples}{transcriptome indices obtained from the Structure Preserving Sampling step.}

\item{default}{logical, when \code{TRUE}, louvain based partition is used and other clustering parameters are ignored.
When \code{FALSE}, hierarchical clustering is used.}

\item{minClusterSize}{integer, specifies the size of the smallest cluster; ignored when \code{default = TRUE}.}

\item{deepSplit}{integer, level of dendrogram split [0-4], higher value produces finer clusters; ignored when \code{default = TRUE}.}

\item{conf}{numeric [0-1], defines the expected confidence of majority for a consensus. Cells remain unassigned when majority is below \code{conf}.}

\item{k_nn}{integers, specifies number of nearest neighbours, defaults to 10.}
}
\value{
List of:\cr
\enumerate{
\item \code{cluster.ident} vector cluster identifiers ranging from 1 to the number of clusters for respective data points.\cr
\item \code{nn.ids}  matrix, each row corresponds to a cell, whose columns depict cluster membership of its neighbours; as returned by the \code{find_ann()} module. \cr
}
Unassigned samples are represented by\code{NA} values.
}
\description{
Performs clustering on sampled cells and Post-hoc Cluster Assignment.
}
\details{
Clustering is carried out in two alternate approaches on the sampled cells.
For the default setting or quick identification of the existing broad clusters,
a louvain based partition is employed. Otherwise for fine-tuned clustering with outliers,
hierarchical clustering is used with \code{cutreeDynamic} for dendrogram cut. Also, Assigns cluster membership to unsampled cells by using cluster membership information of the nearest neighbours.
An approximate nearest neighbour graph is constructed out of the samples population using the \code{find_ann()} module.
Some cells are left un assigned when its neighbour's cluster membership doesn't form a majority as specified by the \code{conf} parameter.
Unassigned cells (\code{NA}) are excluded in the plot or further downstream analysis.
}
