% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{Cluster}
\alias{Cluster}
\title{Clustering of samples}
\usage{
Cluster(object, use.subsamples = TRUE, method = "default",
  use.reduced.dims = NULL, minClusterSize = 20, deepSplit = 3,
  k_nn = 10, conf = 0.75, use.previous = FALSE, ...)
}
\arguments{
\item{object}{A SingleCellExperiment object containing normalized expression values in \code{"normcounts"}.}

\item{use.subsamples}{uses the indices obtained from the sampling method.}

\item{method}{character, one of c("default", "hclust", "kmeans").,
In the default mode, louvain based partition is used.
When \code{hclust}, hierarchical clustering is used.}

\item{use.reduced.dims}{optional, when used, the name of the \code{reducedDim()} table to be used for clustering.}

\item{minClusterSize}{integer, specifies the size of the smallest cluster; works when \code{method = "hclust"}.}

\item{deepSplit}{integer, level of dendrogram split [0-4], higher value produces finer clusters; ignored when
\code{method = "hclust"}.}

\item{k_nn}{integers, specifies number of nearest neighbours, defaults to 10.}

\item{conf}{numeric [0-1], defines the expected confidence of majority for a consensus. Cells remain unassigned when majority is below \code{conf}.}

\item{use.previous}{optional, when \code{TRUE}, the clustering step is skipped and only
the post-hoc clusering is repeated with the new \code{conf}.}

\item{...}{For the \code{kmeans} method option, the argument \code{center} must be passed which specifies the number of clusters.}
}
\value{
List of:\cr
\enumerate{
\item \code{cluster.ident} vector cluster identifiers ranging from 1 to the number of clusters for respective data points.\cr
\item \code{nn.ids}  matrix, each row corresponds to a cell, whose columns depict cluster membership of its neighbours; as returned by the \code{find_ann()} module. \cr
}

Unassigned samples are represented by\code{NA} values.
}
\description{
Performs clustering on sampled cells and Post-hoc Cluster Assignment.
}
\details{
Clustering is carried out in two alternate approaches on the sampled cells.
For the default setting or quick identification of the existing broad clusters,
a Louvain based partition is employed. Otherwise for fine-tuned clustering with outliers,
hierarchical clustering is used with \code{cutreeDynamic} for dendrogram cut. Also, Assigns cluster membership to unsampled cells by using cluster membership information of the nearest neighbours.
An approximate nearest neighbour graph is constructed out of the samples population using the \code{find_ann()} module.
Some cells are left un-assigned when its neighbour's cluster membership doesn't form a majority as specified by the \code{conf} parameter.
Unassigned cells (\code{NA}) are excluded in the plot or further downstream analysis.
}
\examples{
library(SingleCellExperiment)
ncells <- 1000
ngenes <- 2000
x <- matrix(rpois(ncells*ngenes, lambda = 10), ncol=ncells, nrow=ngenes, byrow=TRUE)
rownames(x) <- paste0("Gene", seq_len(ngenes))
colnames(x) <- paste0("Cell", seq_len(ncells))
sce <- SingleCellExperiment(list(counts=x))
sce <- CountNormalize(sce)
sce <- RankGenes(sce)
sce <- Cluster(sce,  use.subsamples=FALSE, conf=0.1)
}
