% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{compute_2d_embedding}
\alias{compute_2d_embedding}
\title{Two Dimensional Embedding for Visualization}
\usage{
compute_2d_embedding(data, sp.samples, clust.list)
}
\arguments{
\item{data}{normalized expression matrix containing the subset of samples and PCA based selected genes.}

\item{sp.samples}{transcriptome indices obtained from the Structure Preserving Sampling step.}

\item{clust.list}{list object as returned by \code{cluster.cells} module.}
}
\value{
numeric two-column matrix containing the 2D embedding of all data points.
}
\description{
2-D embedding of transcriptomes for visualization
}
\details{
Perms t-SNE for 2D embedding on sampled cells followed by the post-hoc projection of 2D co-ordinated of unsampled cells.
An ad-hoc clean up is also carried out to keep away stray points away from the plot - NA value is assigned to such points.
}
