% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{dispersion_genes}
\alias{dispersion_genes}
\title{Top Dispersed Genes}
\usage{
dispersion_genes(normalized_data, ngenes_keep = 1000)
}
\arguments{
\item{normalized_data}{object reurned by \code{normalize()} module.}

\item{ngenes_keep}{integer to return top ranking \code{ngenes_keep} number of genes.}
}
\value{
vector of gene names of length \code{ngenes_keep} .
}
\description{
Get variable genes from normalized UMI counts using Fano Factor metric.
}
\details{
Compute Fano Factor metric for each gene. The metric computes the median absolute deviation of dispersion across multiple bins for each gene.
}
