% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{.annPartition}
\alias{.annPartition}
\title{ANN Graph Partition}
\usage{
.annPartition(data)
}
\arguments{
\item{data}{numeric matrix with \code{n} samples in rows}
}
\value{
two column integer matrix, first column represents sample id, second column contain cluster membership id
}
\description{
ANN Graph Partition
}
