% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropclust.R
\docType{package}
\name{dropClust}
\alias{dropClust}
\alias{dropClust-package}
\title{dropClust}
\description{
Contains functions for Single Cell RNA Seq cluster analysis.
}
\author{
Debajyoti Sinha \email{deb.sinha@hotmail.com}
}
