% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{filter_cells}
\alias{filter_cells}
\title{Cell Filter}
\usage{
filter_cells(data, min_count = 3, ql_th = 0.001, qh_th = 1)
}
\arguments{
\item{data}{named list containing 3 objects: \cr
\enumerate{
\item \code{mat} sparse numeric matrix with count expression data, with genes in columns. \cr
\item \code{gene_symbols} string vector with unique gene names for each column in \code{mat}. \cr
\item \code{barcodes} string vector of length equal to number of rows in \code{mat} identifying each transcriptome per row. \cr
}}

\item{min_count}{integer threshold for expression count}

\item{ql_th}{quantile at probability with values in [0,1] for lower limit}

\item{qh_th}{quantile at probability with values in [0,1] for upper limit}
}
\value{
updated input \code{data} list appended with character vector named \code{keep_cells} containing indices of good cells.
}
\description{
Filter poor quality cells.
}
\details{
Keep only those cells expressing atleast count = \code{min_count} in the number of genes specified within the quantile range between \code{ql_th} and \code{qh_th}.
}
