% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{FilterCells}
\alias{FilterCells}
\title{Filter out poor quality cells}
\usage{
FilterCells(object, min_count = 3, ql_th = 0.001, qh_th = 1)
}
\arguments{
\item{object}{the SingleCellExperiment object to filter}

\item{min_count}{integer threshold for expression count}

\item{ql_th}{quantile at probability with values in [0,1] for lower limit}

\item{qh_th}{quantile at probability with values in [0,1] for upper limit}
}
\value{
SingleCellExperiment object with the bad cells removed
}
\description{
Keep only those cells (columns) expressing atleast count = \code{min_count} in the number of genes specified within the quantile range between \code{ql_th} and \code{qh_th}.
}
\examples{
library(SingleCellExperiment)
counts <- matrix(rpois(100, lambda = 10), ncol=10, nrow=10)
object <- SingleCellExperiment(assays = list(counts = counts))
FilterCells(object)
}
