% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{FilterGenes}
\alias{FilterGenes}
\title{Filter out poor quality genes}
\usage{
FilterGenes(object, min_count = 2, min_cell = 3)
}
\arguments{
\item{object}{the SingleCellExperiment  to filter genes}

\item{min_count}{integer threshold for expression count}

\item{min_cell}{integer threshold for number of cells expressing a particular gene}
}
\value{
SingleCellExperiment object with the bad genes removed
}
\description{
Genes (rows) with count greater than \code{min_count} and not expressed in atleast \code{min_cell} cells are removed.
}
\examples{
library(SingleCellExperiment)
counts <- matrix(rpois(100, lambda = 10), ncol=10, nrow=10)
object <- SingleCellExperiment(assays = list(counts = counts))
FilterGenes(object)
}
