% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{find_ann}
\alias{find_ann}
\title{ANN Search}
\usage{
find_ann(data, idx, n_nn = 10)
}
\arguments{
\item{data}{numeric matrix with \code{n} samples in rows}

\item{idx}{integers to specify specific sample indices. The ANN graph is fit on these samples only.}

\item{n_nn}{integers, specifies number of nearest neighbours.}
}
\value{
integer matrix of dimension \code{n x n_nn}
}
\description{
ANN Search
}
