% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_genes.R
\name{FindMarkers}
\alias{FindMarkers}
\title{Find DE genes}
\usage{
FindMarkers(object, selected_clusters = NA, lfc_th, q_th, nDE = 30)
}
\arguments{
\item{object}{A SingleCellExperiment object containing normalized expression values in \code{"normcounts"}.}

\item{selected_clusters}{vector of selected cluster identifier to be considered. When unspecified (=NA), defaults to all predicted clusters.}

\item{lfc_th}{numeric, [0,1] log2 fold change threshold value to define significance.}

\item{q_th}{numeric, [0,1] fdr adjusted p-value to define significance.}

\item{nDE}{integer, specifies the number of DE genes to return per cluster default = 30.}
}
\value{
list containing:\cr
\itemize{
\item \code{DE_up} vector of unique DE gene names\cr
\item \code{DE_res} list of data frames for respective cluster specific DE genes, each row of a data frame mentions the gene significant pvalues, qvalues and LFC values}
}
\description{
Find cluster specific differentially expressed genes
}
\details{
Performs significance tests in one-vs-all manner to determine cluster specific genes: two-sample Wilcoxon ('Mann-Whitney') test followed by 'fdr' adjustment and log-fold-change.
}
