% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getColors}
\alias{getColors}
\title{Select custom colors}
\usage{
getColors(n)
}
\arguments{
\item{n}{number of colors to generate.}
}
\value{
vector of colour hexcodes of length n.
}
\description{
Produces custom colors.
}
