% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{initial.samples}
\alias{initial.samples}
\title{Initial 1/3rd sampling if very big dataset}
\usage{
initial.samples(data_list, rare_data = NULL)
}
\arguments{
\item{data_list}{list containing \code{read10x()} like structure}

\item{rare_data}{optional, return object obtained from the \code{get_rare_genes()} module.}
}
\value{
two column integer matrix, first column represents sample id, second column contain cluster membership id
}
\description{
Initial 1/3rd sampling if very big dataset
}
