% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{Log2Normalize}
\alias{Log2Normalize}
\title{Compute log normalized expression values from count data in a SingleCellExperiment object.}
\usage{
Log2Normalize(x, return.sparse = FALSE)
}
\arguments{
\item{x}{either a sparse or a dense matrix.}

\item{return.sparse}{Logical when \code{TRUE} return type is same as input, when
\code{FLASE} a dense matrix is returned.}
}
\value{
log normalized matrix same as the input dimensions.
}
\description{
Compute log normalized expression values from count data in a SingleCellExperiment object.
}
\details{
The input matrix is transformed into the log base 2 scale after addition of a pseudo count \code{1}.
}
\examples{
a = matrix(rpois(100, lambda = 10), ncol=10, nrow=10)

lnorm <- Log2Normalize(a)
}
