% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{matrix.transform}
\alias{matrix.transform}
\title{Subset Normalized Matrix}
\usage{
matrix.transform(normalized_data, dp_genes, rare_data = NULL,
  batch_correction = FALSE)
}
\arguments{
\item{normalized_data}{object returned by \code{normalize()} module.}

\item{dp_genes}{gene names returned by \code{dispersion_genes()} module.}

\item{rare_data}{optional, return object obtained from the \code{get_rare_genes()} module.}

\item{batch_correction}{logical, when \code{TRUE} performs batch correction with default parameters.}
}
\value{
numeric matrix.
}
\description{
Subset Normalized matrix to include highly dispersed genes and suspected rare cell genes.
}
\details{
Subset input normalized matrix to include only \code{dp_genes} and \code{rare_genes}.
}
