% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{normalize}
\alias{normalize}
\title{UMI Count Normalization with Gene Filter}
\usage{
normalize(x, min.count = 2, min.cell = 3)
}
\arguments{
\item{x}{named list containing 3 objects: \cr
\enumerate{
\item \code{mat} sparse numeric matrix with count expression data, with genes in columns. \cr
\item \code{gene_symbols} string vector with unique gene names for each column in \code{mat}. \cr
\item \code{barcodes} string vector of length equal to number of rows in \code{mat} idendifying each transcriptome per row. \cr
}}

\item{min.count}{integer threshold for expression count}

\item{min.cell}{integer threshold for number of cells expressing a particular gene}
}
\value{
list containing normalized matrix \code{m} and a character vector named \code{use_genes} of length = dim(m)[2] containing the gene names .
}
\description{
Matrix normalization by UMI count, also remove poorly expressed genes.
}
\details{
Normalizes count value using: \code{total count/median} of respective cells.
Genes with count greater than \code{min.count} and not expressed in atleast \code{min.cell} cells are removed.
}
