% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{pc_genes}
\alias{pc_genes}
\title{Select PCA based Genes}
\usage{
pc_genes(mat, top = 200)
}
\arguments{
\item{mat}{normalized expression matrix containing the subset of samples. Subsetting the return matrix by \code{matrix.subset()} module with samples from primary clusters.}

\item{top}{integer specifying to number of genes to return in their order of ranking}
}
\value{
integer vector containing the indices of \code{top} number of genes in the input matrix
}
\description{
Performs gene selection on sampled cells based on PCA loadings
}
\details{
Genes are ranked for selection in 3 steps:\cr
\enumerate{
\item First 50 principal components are obtained using Singular value Decomposition is used as implemented in the \code{irlba} R package.
\item Among the first 50 components, top 10 components are selected in the order of their modality.
\item Genes are ordered based on their average loadings in the rotation matrix containing the top 10 components.
}
}
