% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Build heatmap object}
\usage{
plot_heatmap(de_data, DE_res, selected_clusters = NA, nDE = 10)
}
\arguments{
\item{de_data}{list containing (1) matrix subset, each row corresponds to a transcriptome sample, columns represent genes; (2) predicted labels of the samples in matrix}

\item{DE_res}{list obtained from the \code{DE_genes} module.}

\item{selected_clusters}{vector of selected clusters to be considered. When unspecified (=NA), defaults to all predicted clusters.}

\item{nDE}{integer, specifies the number of DE genes to appear in the heatmap.}
}
\value{
heatmap object.
}
\description{
Produces custom heatmap uses \code{pheatmap::pheatmap()}.
}
