% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_markers}
\alias{plot_markers}
\title{Violin Plot}
\usage{
plot_markers(de_data, gene.list, filename = NA)
}
\arguments{
\item{de_data}{list containing (1) matrix subset, each row corresponds to a transcriptome sample, columns represent genes; (2) predicted labels of the samples in matrix}

\item{gene.list}{vector of specific marker genes which is a subset of \code{data} column names.}

\item{filename}{character specifying file location to save plot.}

\item{cluster.id}{vector of cluster identifiers of length same as the number of rows in \code{data}.}
}
\value{
list of ggplot objects
}
\description{
Produces violin plots for selected marker genes across predicted groups pf cells
}
