% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emeddings.R
\name{PlotEmbedding}
\alias{PlotEmbedding}
\title{Two Dimensional Embedding for Visualization}
\usage{
PlotEmbedding(object, embedding = "tsne", use.subsamples = TRUE, ...)
}
\arguments{
\item{object}{A SingleCellExperiment object containing normalized expression values in \code{"normcounts"}.}

\item{embedding}{Type of embedding to use, one of \code{c("tsne", "umap")}. defaults to "tsne".
The \code{uwot} implementation of \code{umap} is used.
Specific arguments may be passed with relevant umap argument names.}

\item{use.subsamples}{logical indicating if sub-samples should be used for the post-hoc projection of
the coordinates of remaining samples.}

\item{...}{\code{umap} arguments may be passed.}
}
\value{
A SingleCellExperiment object with a new entry under the
\code{reducedDim()} containter to store the reduced dimension components
with the name from the argument \code{embedding}.
}
\description{
2-D embedding of transcriptomes for visualization
}
\details{
Performs 2D embedding on sampled cells followed by the post-hoc projection of 2D co-ordinated of unsampled cells.
An ad-hoc clean up is also carried out to keep away stray points away from the plot - NA value is assigned to such points.
}
\examples{
library(SingleCellExperiment)
ncells <- 100
ngenes <- 2000
x <- matrix(rpois(ncells*ngenes, lambda = 10), ncol=ncells, nrow=ngenes, byrow=TRUE)
rownames(x) <- paste0("Gene", seq_len(ngenes))
colnames(x) <- paste0("Cell", seq_len(ncells))
sce <- SingleCellExperiment(list(counts=x))
sce <- CountNormalize(sce)
sce <- RankGenes(sce)
sce <- PlotEmbedding(sce)
plot(reducedDim(sce, "tsne"))
}
