% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PlotHeatmap}
\alias{PlotHeatmap}
\title{Build heatmap object}
\usage{
PlotHeatmap(object, DE_res, selected_clusters = NA, nDE = 10)
}
\arguments{
\item{object}{SingleCellExperiment object}

\item{DE_res}{list obtained from the \code{FindMarkers} module.}

\item{selected_clusters}{vector of selected clusters to be considered. When unspecified (=NA), defaults to all predicted clusters.}

\item{nDE}{integer, specifies the number of DE genes to appear in the heatmap.}
}
\value{
heatmap object.
}
\description{
Produces custom heatmap uses \code{pheatmap::pheatmap()}.
}
