% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PlotMarkers}
\alias{PlotMarkers}
\title{Violin Plot}
\usage{
PlotMarkers(object, gene.list, filename = NA)
}
\arguments{
\item{object}{SingleCellExperiment object}

\item{gene.list}{vector of specific marker genes which is a subset of \code{data} column names.}

\item{filename}{character specifying file location to save plot.}
}
\value{
list of ggplot objects
}
\description{
Produces violin plots for selected marker genes across predicted groups pf cells
}
