% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvgs.R
\name{RankGenes}
\alias{RankGenes}
\title{Find highly variable genes}
\usage{
RankGenes(object, ngenes_keep = 1000)
}
\arguments{
\item{object}{A SingleCellExperiment object containing normalized expression values in \code{"normcounts"}.}

\item{ngenes_keep}{integer to return top ranking \code{ngenes_keep} number of genes.}
}
\value{
A SingleCellExperiment object with an additional column named \code{HVG} in \code{rowData} column.
The column stores a a logical value against each gene to indicate if it has been ranked within the top
\code{ngenes_keep}. It also generates an additional column \code{dispersion_norm} in \code{rowData} to
store the dispersion metric against each gene.
}
\description{
Get variable genes from normalized UMI counts using Fano Factor metric.
}
\details{
Compute Fano Factor metric for each gene. The metric computes the median absolute deviation of dispersion across multiple bins for each gene.
}
\examples{
library(SingleCellExperiment)
ncells <- 100
ngenes <- 2000
x <- matrix(rpois(ncells*ngenes, lambda = 10), ncol=ncells, nrow=ngenes, byrow=TRUE)
rownames(x) <- paste0("Gene", seq_len(ngenes))
colnames(x) <- paste0("Cell", seq_len(ncells))
sce <- SingleCellExperiment(list(counts=x))
sce <- CountNormalize(sce)
sce <- RankGenes(sce)
}
