% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvgs.R
\name{RankPCAGenes}
\alias{RankPCAGenes}
\title{Select PCA based Genes}
\usage{
RankPCAGenes(object, top = 200)
}
\arguments{
\item{object}{A SingleCellExperiment object containing normalized expression values in \code{"normcounts"}.}

\item{top}{integer specifying to number of genes to return in their order of ranking}
}
\value{
A SingleCellExperiment object with an additional column named \code{PCAGenes} in \code{rowData} column.
The column stores a a logical value against each gene to indicate if it has been ranked within the \code{top}.
}
\description{
Performs gene selection on sampled cells based on PCA loadings
}
\details{
Genes are ranked for selection in 3 steps:\cr
\enumerate{
\item First 50 principal components are obtained using Singular value Decomposition is used as implemented in the \code{irlba} R package.
\item Among the first 50 components, top 10 components are selected in the order of their modality.
\item Genes are ordered based on their average loadings in the rotation matrix containing the top 10 components.
}
}
\examples{
library(SingleCellExperiment)
ncells <- 1000
ngenes <- 2000
x <- matrix(rpois(ncells*ngenes, lambda = 10), ncol=ncells, nrow=ngenes, byrow=TRUE)
rownames(x) <- paste0("Gene", seq_len(ngenes))
colnames(x) <- paste0("Cell", seq_len(ncells))
sce <- SingleCellExperiment(list(counts=x))
sce <- CountNormalize(sce)
sce <- RankGenes(sce)
sce <- Sampling(sce)
sce <- RankPCAGenes(sce)
}
