% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{read_data}
\alias{read_data}
\title{Read data}
\usage{
read_data(data.path, format = "10X", header = TRUE, sep = ",",
  quote = "\\"")
}
\arguments{
\item{data.path}{When format="10X", specify directory location containing 10X files: \cr
\enumerate{
\item matrix.mtx, sparse matrix format file
\item genes.tsv, gene names tab-separated file
\item barcodes.tsv, barcodes tab-separated file
}; with format="txt", specify the location of the csv matrix file.}

\item{format}{One of c("10X", "txt"). Defaults to "10X".}

\item{header}{TRUE/FALSE. Accept the header row to contain gene names within the csv matrix, defaults to TRUE. When FALSE, columns are labelled automatically.}

\item{sep}{mention the character separator, defaults to ','.}
}
\value{
list containing the 3 input files.
}
\description{
Read 10X or csv data,
}
\details{
Loads expression data from 10X format directory or csv file as specified. The rows in the matrix are considered genes; the columns represent samples
}
