% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{sc_metric}
\alias{sc_metric}
\title{Cluster Validity Index}
\usage{
sc_metric(predicted_clusters, known_types, show_table = TRUE)
}
\arguments{
\item{predicted_clusters}{vector of predicted cluster IDs.}

\item{known_types}{vector of annotated IDs corresponding to \code{predicted_clusters} in the same order.}

\item{show_table}{logical, When TRUE, shows the contingency table of concordance between the IDs.}
}
\value{
numeric a vector of three metrics.
}
\description{
Validate clusters with known annotations when available.
}
\details{
Computes cluster external validity metrics in terms of Rand Index, Adjusted Rand Index and Purity. \cr
Purity of a cluster is measured as a recall of the largest known cell type within a predicted cluster. Total purity across all clusters is returned.
}
