% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ScatterPlot}
\alias{ScatterPlot}
\title{Plotting cells}
\usage{
ScatterPlot(object, labels = NULL, filename = NA, title)
}
\arguments{
\item{object}{A SingleCellExperiment object after performing clustering and plot embedding.}

\item{labels}{A vector of length equal to the number of cells matching the object assay columns.
The label could be cluster identities or other annotations correcponding to each cell in the same order as appearing in the assay column-wiseS.}

\item{filename}{(optional) specify file path to save plot in pdf format.}

\item{title}{character, specify plot title.}
}
\value{
grob object
}
\description{
Scatter plot of cells in two dimensions
}
