% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getColors}
\alias{getColors}
\title{Select custom colours}
\usage{
getColors(n)
}
\arguments{
\item{n}{number of colours to generate.}
}
\value{
vector of colour hex codes of length n.
}
\description{
Produces custom colours.
}
