% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ScatterPlot}
\alias{ScatterPlot}
\title{Plotting cells}
\usage{
ScatterPlot(data, filename = NA, title)
}
\arguments{
\item{data}{data frame containing 3 columns:\cr
\enumerate{
    \item \code{Y1}: numeric values of each point in x-axis.
    \item \code{Y2}: numeric values of each point in y-axis.
    \item \code{color}: factor, cluster identifiers for each point.
}}

\item{filename}{(optional) specify file path to save plot in pdf format.}

\item{title}{character, specify plot title.}
}
\value{
grob object
}
\description{
Scatter plot of cells in two dimensions
}
